The About Page
I made a super simple about page (same template as home page), there seems to be an issue when I try to link the 'About' part of the main menu to about.html so if someone can do that that'd be great. Also feel free to change the styling/check over everything before merging to main.